refactor navigation subsystem
