+++ apps/holtburger-cli/src/pages/game/domains/combat.rs
@@ -1,4 +1,5 @@
 use super::*;
+use super::interaction;
 use crate::pages::game::combat as combat_model;
 use holtburger_core::ActionResultReason;
 
@@ -132,10 +133,10 @@ fn start_explicit_attack(state: &mut GameState, target_guid: Guid) -> UpdateResu
         state.view.active_interaction,
         Some(Interaction::Approaching { .. }) | Some(Interaction::Following { .. })
     ) {
-        super::navigation::clear_active_interaction(state, &mut result);
+        interaction::clear_active_interaction(state, &mut result);
     }
 
-    super::navigation::set_active_interaction(
+    interaction::set_active_interaction(
         state,
         Some(Interaction::Targeting { target_guid }),
         &mut result,
