+++ apps/holtburger-cli/src/pages/game/domains/inventory.rs
@@ -1,5 +1,5 @@
 use super::context;
-use super::navigation;
+use super::interaction;
 use super::*;
 use holtburger_core::client::controllers::Controller;
 
@@ -30,7 +30,7 @@ pub(super) fn reduce_action(state: &mut GameState, action: AppAction) -> UpdateR
                     .retain(|queued_guid| *queued_guid != guid);
 
                 if session.queued_items.is_empty() {
-                    navigation::clear_active_interaction(state, &mut result);
+                    interaction::clear_active_interaction(state, &mut result);
                     state.view.salvaging = None;
                 }
                 result.request_redraw(RedrawPriority::Immediate);
@@ -46,7 +46,7 @@ pub(super) fn reduce_action(state: &mut GameState, action: AppAction) -> UpdateR
                     items: item_guids,
                 });
             }
-            navigation::clear_active_interaction(state, &mut result);
+            interaction::clear_active_interaction(state, &mut result);
             state.view.salvaging = None;
             result.request_redraw(RedrawPriority::Immediate);
         }
@@ -93,7 +93,7 @@ pub(super) fn reduce_action(state: &mut GameState, action: AppAction) -> UpdateR
                 container,
                 placement: 0,
             });
-            navigation::clear_active_interaction(state, &mut result);
+            interaction::clear_active_interaction(state, &mut result);
         }
         AppAction::StackItems {
             source,
@@ -279,7 +279,7 @@ pub(super) fn handle_entity_removed(state: &mut GameState, guid: Guid) -> Update
         state.view.active_interaction,
         Some(Interaction::Targeting { target_guid }) if target_guid == guid
     ) {
-        navigation::clear_active_interaction(state, &mut result);
+        interaction::clear_active_interaction(state, &mut result);
     }
     if let Some(session) = state.view.salvaging.as_mut() {
         session
@@ -288,7 +288,7 @@ pub(super) fn handle_entity_removed(state: &mut GameState, guid: Guid) -> Update
         if session.ust_guid == guid {
             state.view.salvaging = None;
             if state.view.active_interaction == Some(Interaction::Salvaging) {
-                navigation::clear_active_interaction(state, &mut result);
+                interaction::clear_active_interaction(state, &mut result);
             }
         }
     }
